requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
matplotlib>=3.7.0
requests>=2.31.0
